+++ lib/drone/application.ex
@@ -5,6 +5,14 @@ defmodule Drone.Application do
 
   @impl true
   def start(_type, _args) do
+    _ =
+      :ets.new(Drone.Swarm.members_table(), [
+        :named_table,
+        :public,
+        :set,
+        read_concurrency: true
+      ])
+
     children = [
       {Registry, keys: :unique, name: Drone.Vehicle.Registry},
       {Registry, keys: :unique, name: Drone.Swarm.Registry},